requirements.txt
ipdb
ipython
jupyter
Django==1.11.6
Pillow==4.3.0
PyYAML==3.12
Theano==0.9.0
docopt==0.6.2
easydict==1.7
enum34==1.1.6
flask==0.12.2
fuel==0.2.0
funcsigs==1.0.2
h5py==2.7.1
image==1.5.16
keras==2.0.8
mock==2.0.0
nltk==3.2.5
numpy==1.13.3
olefile==0.44
pandas==0.20.3
pbr==3.1.1
prettytensor==0.7.3
progressbar==2.3
protobuf==3.4.0
python-dateutil==2.6.1
pytz==2017.2
scikit-image==0.13.1
scipy==0.19.1
six==1.11.0
torchfile==0.1.0
